Attendance/requirements.txt
telebot
timedelta
geopy
psycopg2-binary
python-dotenv
persiantools
